PAR-56/61 - Splits drag/sort function into multiple modules, removes duplicate line creation in svg service
