+++ src/app/classes/diagram/draggable.ts
@@ -1,0 +1,6 @@
+export type Draggable = {
+    transition: HTMLElement;
+    infoElement?: HTMLElement;
+    incomingArcs: Array<HTMLElement>;
+    outgoingArcs: Array<HTMLElement>;
+};
