+++ src/app/services/moving/dragging-helper.fn.ts
@@ -1,0 +1,18 @@
+export function asInt(elem: Element, attribute: string): number {
+    return parseInt(elem.getAttribute(attribute) ?? '0');
+}
+
+export function getYAttribute(element: HTMLElement): string {
+    return element.nodeName === 'rect' ? 'y' : 'cy';
+}
+
+export function getXAttribute(element: HTMLElement): string {
+    return element.nodeName === 'rect' ? 'x' : 'cx';
+}
+
+export function getAttributePraefix(e: HTMLElement): string {
+    if (e.nodeName === 'circle') {
+        return 'c';
+    }
+    return '';
+}
