+++ src/app/services/moving/factory/draggable-creation.service.spec.ts
@@ -1,0 +1,18 @@
+import { TestBed } from '@angular/core/testing';
+
+import { DraggingCreationService } from './draggable-creation.service';
+
+describe('ParserService', () => {
+    let service: DraggingCreationService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [{ provide: DraggingCreationService, useValue: {} }],
+        });
+        service = TestBed.inject(DraggingCreationService);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});
