+++ src/app/services/moving/find/find-elements.service.spec.ts
@@ -1,0 +1,18 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FindElementsService } from './find-elements.service';
+
+describe('ParserService', () => {
+    let service: FindElementsService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [{ provide: FindElementsService, useValue: {} }],
+        });
+        service = TestBed.inject(FindElementsService);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});
