+++ src/app/services/moving/move/move-elements.service.spec.ts
@@ -1,0 +1,18 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MoveElementsService } from './move-elements.service';
+
+describe('ParserService', () => {
+    let service: MoveElementsService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [{ provide: MoveElementsService, useValue: {} }],
+        });
+        service = TestBed.inject(MoveElementsService);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});
