+++ src/app/services/moving/statehandler/statehandler.service.spec.ts
@@ -1,0 +1,18 @@
+import { TestBed } from '@angular/core/testing';
+
+import { StatehandlerService } from './statehandler.service';
+
+describe('ParserService', () => {
+    let service: StatehandlerService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            providers: [{ provide: StatehandlerService, useValue: {} }],
+        });
+        service = TestBed.inject(StatehandlerService);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});
